Huge committ !   Please use the restore tools to correct the database Please be aware, to use the new plugin functions, you must install first the paymentplugin bundle of francesco, maybe he makes an updated version,. delete the data  in plugins/vmpayment/ and reload them via svn. - modelfunctions small fix prevents to add 0 data - parameterparser, the function  list_available_classes (of payment plugins) uses now the installed plugins and determines them by group. - Made an abstract class vmpaymentplugin.php for the new paymentmethod plugin system, working joomla conform, with hooks in the cart. - adjusted the payment model and table and views to the new payment plugin system, big difference is now that the plugin itself saves the plugin type. Therefore the payment type stuff is not needed anylonger. - added hooks to the cart.php - Corrected 2 plugins as exampel, cashondel and authorize - adjusted all xml files of the payment plugins to fit into the new system - adjusted install.sql to new payment table - adjusted user_info table class, removed bank stuff which is now a new paymentplugin - deleted old payment plugin folder
git-svn-id: https://dev.virtuemart.net/svn/virtuemart/trunk/virtuemart@2521 87cc427b-4a1e-0410-b729-b2260ec0f063

+++ administrator/components/com_virtuemart/helpers/modelfunctions.php
@@ -51,6 +51,7 @@ class modelfunctions{
 	 * @param $data data array
 	 */
 	function storeArrayData($table,$fieldId,$fieldData,$id,$data){
+		if(!$data) return false;
 		$db = JFactory::getDBO();
 		$q  = 'DELETE FROM `'.$table.'` WHERE `'.$fieldId.'` = "'.$id.'" ';
 		$db->setQuery($q);
@@ -66,6 +67,7 @@ class modelfunctions{
 			$db->setQuery($q); 
 			$db->query();
 		}
+		return true;
 	}
 	
 	/**
